+++ mpp/src/CircleBatch.cpp
@@ -85,7 +85,7 @@ namespace mpp
 		data.resize(count * vertexSize);
 
 		uint32_t* ptr = (uint32_t*)&data[start * vertexSize]; // Indices will be 16 or 32-bit, so use 32 to cover both
-		int indexBytes = mIndexWidth / 8;
+		int indexBytes = (int)mIndexWidth / 8;
 
 		for (uint32_t i = start; i < count; ++i)
 		{
@@ -134,7 +134,7 @@ namespace mpp
 	{
 		for (size_t i = 0; i < getCapacity(); ++i)
 		{
-			auto x = i * 4;
+			auto x = (uint32_t)i * 4;
 			ms->addTriangle(meshIndex, x + 0, x + 1, x + 2);
 			ms->addTriangle(meshIndex, x + 2, x + 3, x + 0);
 		}
@@ -150,7 +150,7 @@ namespace mpp
 
 	int CircleBatch::getIndexWidth() const
 	{
-		return mIndexWidth;
+		return (int)mIndexWidth;
 	}
 
 	size_t CircleBatch::getPrimitiveCount(size_t objectCount) const
